refactor: move test classes into submodule
