+++ frappe/tests/classes/__init__.py
@@ -1,1 +1,3 @@
-
+from .integration_test_case import IntegrationTestCase
+from .mocked_request_test_case import MockedRequestTestCase
+from .unit_test_case import UnitTestCase
