+++ frappe/tests/utils/__init__.py
@@ -1,430 +1,11 @@
-import copy
-import datetime
-import json
-import os
-import unittest
-from collections.abc import Sequence
-from contextlib import AbstractContextManager, contextmanager
-from pathlib import Path
-from typing import Any
-from unittest.mock import patch
-
-import pytz
+import logging
 
 import frappe
-from frappe.model.base_document import BaseDocument, get_controller
-from frappe.utils import cint
-from frappe.utils.data import get_datetime, get_system_timezone
-
-datetime_like_types = (datetime.datetime, datetime.date, datetime.time, datetime.timedelta)
-import logging
+from frappe.model.base_document import get_controller
 
 logger = logging.Logger(__file__)
 
-# TODO: move to dumpster (backwars compat) - not meant to be a public interface anymore
-from frappe.tests.classes.context_managers import debug_on, timeout
-
 from .generators import *
-
-
-class UnitTestCase(unittest.TestCase):
-	"""Unit test class for Frappe tests.
-
-	This class extends unittest.TestCase and provides additional utilities
-	specific to Frappe framework. It's designed for testing individual
-	components or functions in isolation.
-
-	Key features:
-	- Custom assertions for Frappe-specific comparisons
-	- Utilities for HTML and SQL normalization
-	- Context managers for user switching and time freezing
-
-	Note: If you override `setUpClass`, make sure to call `super().setUpClass()`
-	to maintain the functionality of this base class.
-	"""
-
-	@classmethod
-	def setUpClass(cls) -> None:
-		super().setUpClass()
-		cls.doctype = cls._get_doctype_from_module()
-		cls.module = frappe.get_module(cls.__module__)
-
-	@classmethod
-	def _get_doctype_from_module(cls):
-		module_path = cls.__module__.split(".")
-		try:
-			doctype_index = module_path.index("doctype")
-			doctype_snake_case = module_path[doctype_index + 1]
-			json_file_path = Path(*module_path[:-1]).joinpath(f"{doctype_snake_case}.json")
-			if json_file_path.is_file():
-				doctype_data = json.loads(json_file_path.read_text())
-				return doctype_data.get("name")
-		except (ValueError, IndexError):
-			# 'doctype' not found in module_path
-			pass
-		return None
-
-	def _apply_debug_decorator(self, exceptions=()):
-		setattr(self, self._testMethodName, debug_on(*exceptions)(getattr(self, self._testMethodName)))
-
-	def assertQueryEqual(self, first: str, second: str) -> None:
-		self.assertEqual(self.normalize_sql(first), self.normalize_sql(second))
-
-	def assertSequenceSubset(self, larger: Sequence, smaller: Sequence, msg: str | None = None) -> None:
-		"""Assert that `expected` is a subset of `actual`."""
-		self.assertTrue(set(smaller).issubset(set(larger)), msg=msg)
-
-	# --- Frappe Framework specific assertions
-	def assertDocumentEqual(self, expected: dict | BaseDocument, actual: BaseDocument) -> None:
-		"""Compare a (partial) expected document with actual Document."""
-
-		if isinstance(expected, BaseDocument):
-			expected = expected.as_dict()
-
-		for field, value in expected.items():
-			if isinstance(value, list):
-				actual_child_docs = actual.get(field)
-				self.assertEqual(len(value), len(actual_child_docs), msg=f"{field} length should be same")
-				for exp_child, actual_child in zip(value, actual_child_docs, strict=False):
-					self.assertDocumentEqual(exp_child, actual_child)
-			else:
-				self._compare_field(value, actual.get(field), actual, field)
-
-	def _compare_field(self, expected: Any, actual: Any, doc: BaseDocument, field: str) -> None:
-		msg = f"{field} should be same."
-
-		if isinstance(expected, float):
-			precision = doc.precision(field)
-			self.assertAlmostEqual(
-				expected, actual, places=precision, msg=f"{field} should be same to {precision} digits"
-			)
-		elif isinstance(expected, bool | int):
-			self.assertEqual(expected, cint(actual), msg=msg)
-		elif isinstance(expected, datetime_like_types) or isinstance(actual, datetime_like_types):
-			self.assertEqual(str(expected), str(actual), msg=msg)
-		else:
-			self.assertEqual(expected, actual, msg=msg)
-
-	def normalize_html(self, code: str) -> str:
-		"""Formats HTML consistently so simple string comparisons can work on them."""
-		from bs4 import BeautifulSoup
-
-		return BeautifulSoup(code, "html.parser").prettify(formatter=None)
-
-	@contextmanager
-	def set_user(self, user: str) -> AbstractContextManager[None]:
-		try:
-			old_user = frappe.session.user
-			frappe.set_user(user)
-			yield
-		finally:
-			frappe.set_user(old_user)
-
-	def normalize_sql(self, query: str) -> str:
-		"""Formats SQL consistently so simple string comparisons can work on them."""
-		import sqlparse
-
-		return sqlparse.format(query.strip(), keyword_case="upper", reindent=True, strip_comments=True)
-
-	@classmethod
-	def enable_safe_exec(cls) -> None:
-		"""Enable safe exec and disable them after test case is completed."""
-		from frappe.installer import update_site_config
-		from frappe.utils.safe_exec import SAFE_EXEC_CONFIG_KEY
-
-		cls._common_conf = os.path.join(frappe.local.sites_path, "common_site_config.json")
-		update_site_config(SAFE_EXEC_CONFIG_KEY, 1, validate=False, site_config_path=cls._common_conf)
-
-		cls.addClassCleanup(
-			lambda: update_site_config(
-				SAFE_EXEC_CONFIG_KEY, 0, validate=False, site_config_path=cls._common_conf
-			)
-		)
-
-	@staticmethod
-	@contextmanager
-	def patch_hooks(overridden_hooks: dict) -> AbstractContextManager[None]:
-		get_hooks = frappe.get_hooks
-
-		def patched_hooks(hook=None, default="_KEEP_DEFAULT_LIST", app_name=None):
-			if hook in overridden_hooks:
-				return overridden_hooks[hook]
-			return get_hooks(hook, default, app_name)
-
-		with patch.object(frappe, "get_hooks", patched_hooks):
-			yield
-
-	@contextmanager
-	def freeze_time(
-		self, time_to_freeze: Any, is_utc: bool = False, *args: Any, **kwargs: Any
-	) -> AbstractContextManager[None]:
-		from freezegun import freeze_time
-
-		if not is_utc:
-			# Freeze time expects UTC or tzaware objects. We have neither, so convert to UTC.
-			timezone = pytz.timezone(get_system_timezone())
-			time_to_freeze = timezone.localize(get_datetime(time_to_freeze)).astimezone(pytz.utc)
-
-		with freeze_time(time_to_freeze, *args, **kwargs):
-			yield
-
-
-class IntegrationTestCase(UnitTestCase):
-	"""Integration test class for Frappe tests.
-
-	Key features:
-	- Automatic database setup and teardown
-	- Utilities for managing database connections
-	- Context managers for query counting and Redis call monitoring
-	- Lazy loading of test record dependencies
-
-	Note: If you override `setUpClass`, make sure to call `super().setUpClass()`
-	to maintain the functionality of this base class.
-	"""
-
-	TEST_SITE = "test_site"
-
-	SHOW_TRANSACTION_COMMIT_WARNINGS = False
-	maxDiff = 10_000  # prints long diffs but useful in CI
-
-	@classmethod
-	def setUpClass(cls) -> None:
-		super().setUpClass()
-
-		# Site initialization
-		cls.TEST_SITE = getattr(frappe.local, "site", None) or cls.TEST_SITE
-		frappe.init(cls.TEST_SITE)
-		cls.ADMIN_PASSWORD = frappe.get_conf(cls.TEST_SITE).admin_password
-
-		cls._primary_connection = frappe.local.db
-		cls._secondary_connection = None
-
-		# Create test record dependencies
-		cls._newly_created_test_records = []
-		if cls.doctype and cls.doctype not in frappe.local.test_objects:
-			cls._newly_created_test_records += make_test_records(cls.doctype)
-		for doctype in getattr(cls.module, "test_dependencies", []):
-			if doctype not in frappe.local.test_objects:
-				cls._newly_created_test_records += make_test_records(doctype)
-
-		# flush changes done so far to avoid flake
-		frappe.db.commit()
-		if cls.SHOW_TRANSACTION_COMMIT_WARNINGS:
-			frappe.db.before_commit.add(_commit_watcher)
-
-		# enqueue teardown actions (executed in LIFO order)
-		cls.addClassCleanup(_restore_thread_locals, copy.deepcopy(frappe.local.flags))
-		cls.addClassCleanup(_rollback_db)
-
-	@classmethod
-	def tearDownClass(cls) -> None:
-		# Add any necessary teardown code here
-		super().tearDownClass()
-
-	def setUp(self) -> None:
-		super().setUp()
-		# Add any per-test setup code here
-
-	def tearDown(self) -> None:
-		# Add any per-test teardown code here
-		super().tearDown()
-
-	@contextmanager
-	def primary_connection(self) -> AbstractContextManager[None]:
-		"""Switch to primary DB connection
-
-		This is used for simulating multiple users performing actions by simulating two DB connections"""
-		try:
-			current_conn = frappe.local.db
-			frappe.local.db = self._primary_connection
-			yield
-		finally:
-			frappe.local.db = current_conn
-
-	@contextmanager
-	def secondary_connection(self) -> AbstractContextManager[None]:
-		"""Switch to secondary DB connection."""
-		if self._secondary_connection is None:
-			frappe.connect()  # get second connection
-			self._secondary_connection = frappe.local.db
-
-		try:
-			current_conn = frappe.local.db
-			frappe.local.db = self._secondary_connection
-			yield
-		finally:
-			frappe.local.db = current_conn
-			self.addCleanup(self._rollback_connections)
-
-	def _rollback_connections(self) -> None:
-		self._primary_connection.rollback()
-		self._secondary_connection.rollback()
-
-	@contextmanager
-	def assertQueryCount(self, count: int) -> AbstractContextManager[None]:
-		queries = []
-
-		def _sql_with_count(*args, **kwargs):
-			ret = orig_sql(*args, **kwargs)
-			queries.append(args[0].last_query)
-			return ret
-
-		try:
-			orig_sql = frappe.db.__class__.sql
-			frappe.db.__class__.sql = _sql_with_count
-			yield
-			self.assertLessEqual(len(queries), count, msg="Queries executed: \n" + "\n\n".join(queries))
-		finally:
-			frappe.db.__class__.sql = orig_sql
-
-	@contextmanager
-	def assertRedisCallCounts(self, count: int) -> AbstractContextManager[None]:
-		commands = []
-
-		def execute_command_and_count(*args, **kwargs):
-			ret = orig_execute(*args, **kwargs)
-			key_len = 2
-			if "H" in args[0]:
-				key_len = 3
-			commands.append((args)[:key_len])
-			return ret
-
-		try:
-			orig_execute = frappe.cache.execute_command
-			frappe.cache.execute_command = execute_command_and_count
-			yield
-			self.assertLessEqual(
-				len(commands), count, msg="commands executed: \n" + "\n".join(str(c) for c in commands)
-			)
-		finally:
-			frappe.cache.execute_command = orig_execute
-
-	@contextmanager
-	def assertRowsRead(self, count: int) -> AbstractContextManager[None]:
-		rows_read = 0
-
-		def _sql_with_count(*args, **kwargs):
-			nonlocal rows_read
-
-			ret = orig_sql(*args, **kwargs)
-			# count of last touched rows as per DB-API 2.0 https://peps.python.org/pep-0249/#rowcount
-			rows_read += cint(frappe.db._cursor.rowcount)
-			return ret
-
-		try:
-			orig_sql = frappe.db.sql
-			frappe.db.sql = _sql_with_count
-			yield
-			self.assertLessEqual(rows_read, count, msg="Queries read more rows than expected")
-		finally:
-			frappe.db.sql = orig_sql
-
-	@contextmanager
-	def switch_site(self, site: str) -> AbstractContextManager[None]:
-		"""Switch connection to different site.
-		Note: Drops current site connection completely."""
-
-		try:
-			old_site = frappe.local.site
-			frappe.init(site, force=True)
-			frappe.connect()
-			yield
-		finally:
-			frappe.init(old_site, force=True)
-			frappe.connect()
-
-	@staticmethod
-	@contextmanager
-	def change_settings(doctype, settings_dict=None, /, commit=False, **settings):
-		"""A context manager to ensure that settings are changed before running
-		function and restored after running it regardless of exceptions occurred.
-		This is useful in tests where you want to make changes in a function but
-		don't retain those changes.
-		import and use as decorator to cover full function or using `with` statement.
-
-		example:
-		@change_settings("Print Settings", {"send_print_as_pdf": 1})
-		def test_case(self):
-		        ...
-
-		@change_settings("Print Settings", send_print_as_pdf=1)
-		def test_case(self):
-		        ...
-		"""
-
-		if settings_dict is None:
-			settings_dict = settings
-
-		try:
-			settings = frappe.get_doc(doctype)
-			# remember setting
-			previous_settings = copy.deepcopy(settings_dict)
-			for key in previous_settings:
-				previous_settings[key] = getattr(settings, key)
-
-			# change setting
-			for key, value in settings_dict.items():
-				setattr(settings, key, value)
-			settings.save(ignore_permissions=True)
-			# singles are cached by default, clear to avoid flake
-			frappe.db.value_cache[settings] = {}
-			if commit:
-				frappe.db.commit()
-			yield  # yield control to calling function
-
-		finally:
-			# restore settings
-			settings = frappe.get_doc(doctype)
-			for key, value in previous_settings.items():
-				setattr(settings, key, value)
-			settings.save(ignore_permissions=True)
-			if commit:
-				frappe.db.commit()
-
-
-# TODO: move to dumpster
-FrappeTestCase = IntegrationTestCase
-change_settings = IntegrationTestCase.change_settings
-patch_hooks = UnitTestCase.patch_hooks
-
-
-class MockedRequestTestCase(IntegrationTestCase):
-	def setUp(self):
-		import responses
-
-		self.responses = responses.RequestsMock()
-		self.responses.start()
-
-		self.addCleanup(self.responses.stop)
-		self.addCleanup(self.responses.reset)
-
-		return super().setUp()
-
-
-def _commit_watcher():
-	import traceback
-
-	logger.warning("Transaction committed during tests.")
-	traceback.print_stack(limit=10)
-
-
-def _rollback_db():
-	frappe.db.value_cache = {}
-	frappe.db.rollback()
-
-
-def _restore_thread_locals(flags):
-	frappe.local.flags = flags
-	frappe.local.error_log = []
-	frappe.local.message_log = []
-	frappe.local.debug_log = []
-	frappe.local.conf = frappe._dict(frappe.get_site_config())
-	frappe.local.cache = {}
-	frappe.local.lang = "en"
-	frappe.local.preload_assets = {"style": [], "script": [], "icons": []}
-
-	if hasattr(frappe.local, "request"):
-		delattr(frappe.local, "request")
 
 
 def check_orpahned_doctypes():
@@ -443,3 +24,13 @@ def check_orpahned_doctypes():
 		frappe.throw(
 			"Following doctypes exist in DB without controller.\n {}".format("\n".join(orpahned_doctypes))
 		)
+
+
+# TODO: move to dumpster (backwars compat) - not meant to be a public interface anymore
+from frappe.tests.classes import IntegrationTestCase, MockedRequestTestCase, UnitTestCase
+from frappe.tests.classes.context_managers import debug_on, timeout
+
+# TODO: move to dumpster
+FrappeTestCase = IntegrationTestCase
+change_settings = IntegrationTestCase.change_settings
+patch_hooks = UnitTestCase.patch_hooks
